Add doxygen comments for generator_rational.

+++ source/notations/twist_vector/notation_tv.c
@@ -54,7 +54,7 @@ note_tv_t *note_tv_tvrev(note_tv_t *tv);
 /*
  *  Documentation in header
  */
-int note_tv_encode(char *str, note_tv_t *twistv)
+uint8_t note_tv_encode(char *str, note_tv_t *twistv)
 {
     /*@@@TODO: the return logic here needs to be inverted when error handing is
      * added*/
@@ -75,7 +75,7 @@ int note_tv_encode(char *str, note_tv_t *twistv)
 /*
  *  Documentation in header
  */
-int note_tv_decode(note_tv_t twistv, char *str)
+uint8_t note_tv_decode(note_tv_t twistv, char *str)
 {
     /*@@@TODO: the return logic here needs to be inverted when error handing is
      * added*/
